Move `ImageCommensPresenter` to production, altogether with localization file
