+++ EssentialFeed/EssentialFeed/Feed Presentation/pt-BR.lproj/ImageComments.strings
