+++ EssentialFeed/EssentialFeedTests/Feed Presentation/ImageCommentsPresenterTests.swift
@@ -5,75 +5,6 @@
 import XCTest
 import EssentialFeed
 
-struct ImageCommentsLoadingViewModel {
-	let isLoading: Bool
-}
-
-protocol ImageCommentsLoadingView {
-	func display(_ viewModel: ImageCommentsLoadingViewModel)
-}
-
-struct ImageCommentsErrorViewModel {
-	let message: String?
-}
-
-protocol ImageCommentsErrorView {
-	func display(_ viewModel: ImageCommentsErrorViewModel)
-}
-
-struct ImageCommentsViewModel {
-	let comments: [ImageComment]
-}
-
-protocol ImageCommentsView {
-	func display(_ viewModel: ImageCommentsViewModel)
-}
-
-class ImageCommentsPresenter {
-	private let commentsView: ImageCommentsView
-	private let loadingView: ImageCommentsLoadingView
-	private let errorView: ImageCommentsErrorView
-	
-	init(commentsView: ImageCommentsView, loadingView: ImageCommentsLoadingView, errorView: ImageCommentsErrorView) {
-		self.commentsView = commentsView
-		self.loadingView = loadingView
-		self.errorView = errorView
-	}
-	
-	func didStartLoadingComments() {
-		errorView.display(ImageCommentsErrorViewModel(message: .none))
-		loadingView.display(ImageCommentsLoadingViewModel(isLoading: true))
-	}
-	
-	func didFinishLoadingComments(with comments: [ImageComment]) {
-		commentsView.display(ImageCommentsViewModel(comments: comments))
-		loadingView.display(ImageCommentsLoadingViewModel(isLoading: false))
-	}
-	
-	func didFinishLoadingFeed(with error: Error) {
-		errorView.display(ImageCommentsErrorViewModel(message: imageCommentsLoadError))
-		loadingView.display(ImageCommentsLoadingViewModel(isLoading: false))
-	}
-	
-	private var imageCommentsLoadError: String {
-		NSLocalizedString(
-			"IMAGE_COMMENTS_VIEW_CONNECTION_ERROR",
-			tableName: ImageCommentsPresenter.tableName,
-			bundle: Bundle(for: ImageCommentsPresenter.self),
-			comment: "Error message displayed when we can't load the image comments from the server")
-	}
-	
-	static var title: String {
-		NSLocalizedString(
-			"IMAGE_COMMENTS_VIEW_TITLE",
-			tableName: tableName,
-			bundle: Bundle(for: ImageCommentsPresenter.self),
-			comment: "Title for the image comments view")
-	}
-	
-	static var tableName: String { "ImageComments" }
-}
-
 class ImageCommentsPresenterTests: XCTestCase {
 	
 	func test_init_doesNotSendMessagesToView() {
